+++ frontend/src/pages/LoanPage.jsx
@@ -1,5 +1,0 @@
-const LoanPage = () => {
-  return <div>LoanPage</div>;
-};
-
-export default LoanPage;